requirements.txt
opencv-python>=4.8.0
pyautogui>=0.9.54
pillow>=10.0.0
numpy>=1.24.0
pygetwindow>=0.0.9
mss>=9.0.0
keyboard>=0.13.5
mouse>=0.7.1
pytesseract>=0.3.10
pywin32>=306
psutil>=5.9.0
matplotlib>=3.5.0